- removed upload_endpoint from Attachments - moved attachment interface from Attachments to AttachmentUploader - removed attribute: false from failedAttachments properties across all 3 components
